+++ R/fitDR-bootstrap.R
@@ -1,4 +1,5 @@
-bootDR <- function(f, bootmethod="param", niter=1001, silent=TRUE)
+bootDR <- function(f, bootmethod="param", niter=1001, silent=TRUE,
+                   parallel = c("no", "snow", "multicore"), ncpus)
 {
   if (!inherits(f, "DR"))
     stop("Use only with 'DR' objects")
@@ -6,47 +7,69 @@ bootDR <- function(f, bootmethod="param", niter=1001, silent=TRUE)
     stop("niter must be an integer above 10")
   bootmethod <- match.arg(bootmethod, c("param", "nonparam"))
   
+  parallel <- match.arg(parallel, c("no", "snow", "multicore"))
+  if (parallel == "multicore" & .Platform$OS.type == "windows")
+  {
+    parallel <- "snow"
+    warning("As the multicore option is not supported on Windows it was replaced by snow")
+  }
+  if ((parallel == "snow" | parallel == "multicore") & missing(ncpus)) 
+    stop("You have to specify the number of available processors to parallelize 
+             the bootstrap")
+  
   #simulate bootstrap data
-  if (bootmethod == "param") { # parametric bootstrap
+  if (bootmethod == "param") 
+  { # parametric bootstrap
     rdistname <- paste("r", f$distname, sep="")
     if (!exists(rdistname, mode="function"))
       stop(paste("The ", rdistname, " function must be defined"))
     rdata <- do.call(rdistname, c(list(n=niter*f$n), as.list(f$estimate)))
     dim(rdata) <- c(f$n, niter)
-  }
-  else { # non parametric bootstrap
+  }else 
+  { # non parametric bootstrap
     rdata <- sample(f$data, size=niter*f$n, replace=TRUE)
     dim(rdata) <- c(f$n, niter)
   }
   
   #compute bootstrap estimates
-  start <- as.list(f$estimate) #a named vector is no longer is accepted as starting values.
-  #if (is.null(f$dots))
-  func <- function(iter) {
-    res <- try(do.call(fitDR, list(x=rdata[, iter], dist=f$distname, start=start)), silent=silent)
+  start <- as.list(f$estimate) #a named vector is no longer accepted as starting values.
+  
+  func <- function(iter) 
+  {
+    res <- try(do.call(fitDR, list(x=rdata[, iter], dist=f$distname, start=start,
+                                   optim.method=f$optim.method, control=f$control)
+                       ), silent=silent)
     
     if(class(res)[1] == "try-error")
       return(c(rep(NA, length(start)), 100))
     else
       return(c(res$estimate, res$convergence))
   }
-  #else
-  #  func <- function(iter) {
-  #    res <- do.call(fitDR, c(list(x=rdata[, iter], dist=f$distname, start=start, fix.arg=f$fix.arg), f$dots))
-  #    return(c(res$estimate, res$convergence))
-  #  }
+  
   owarn <- getOption("warn")
   oerr <- getOption("show.error.messages")
-  #print(owarn)
-  #print(oerr)
   options(warn=ifelse(silent, -1, 0), show.error.messages=!silent)
-  resboot <- sapply(1:niter, func)
+  
+  # parallel or sequential computation
+  if (parallel != "no") 
+  {
+    if (parallel == "snow") type <- "PSOCK"
+    else if (parallel == "multicore") type <- "FORK"
+    clus <- parallel::makeCluster(ncpus, type = type)
+    parallel::clusterEvalQ(clus, {library(mbbefd)})
+    resboot <- parallel::parSapply(clus, 1:niter, func)
+    parallel::stopCluster(clus)
+  }
+  else
+  {
+    resboot <- sapply(1:niter, func)
+  }
+  
   options(warn=owarn, show.error.messages=oerr)
-  #print(owarn)
-  #print(oerr)
   
   rownames(resboot) <- c(names(f$estimate), "convergence")
-  if (length(resboot[, 1])>2) {
+  if (length(resboot[, 1])>2) 
+  {
     estim <- data.frame(t(resboot)[, -length(resboot[, 1])])
     bootCI <- cbind(apply(resboot[-length(resboot[, 1]), ], 1, median, na.rm=TRUE), 
                     apply(resboot[-length(resboot[, 1]), ], 1, quantile, 0.025, na.rm=TRUE),