+++ tests/test-fit-mbbefdab.R
@@ -1,10 +1,10 @@
 library(mbbefd)
 library(fitdistrplus)
+set.seed(123456)
 
 n <- 1e3
 nboot <- 1000
 nboot <- 10
-set.seed(123456)
 lossrate <- rmbbefd(n, 1/2, 1/10)
 
 
@@ -30,15 +30,18 @@ vcov(f1)
 # points(1/2, 1/10, pch="x", col="black")
 
 b1 <- bootDR(f1, niter=nboot, silent=TRUE)
-plot(b1, enhance=TRUE, trueval=c(1/2, 1/10))
 
+if(sum(b1$converg == 0) > 2)
+{
+  plot(b1, enhance=TRUE, trueval=c(1/2, 1/10))
+  plot(density(b1))
+}
 
 f2 <- fitDR(lossrate, "mbbefd", method="tlmme")
 summary(f2)
 
 
 
-set.seed(123456)
 lossrate <- rmbbefd(n, -1/2, 5)
 
 
@@ -46,5 +49,9 @@ f1 <- fitDR(lossrate, "mbbefd")
 summary(f1)
 
 b1 <- bootDR(f1, niter=nboot, silent=TRUE)
-plot(b1, enhance=TRUE, trueval=c(-1/2, 5))
+if(sum(b1$converg == 0) > 2)
+{
+  plot(b1, enhance=TRUE, trueval=c(-1/2, 5))
+  plot(density(b1))
+}