+++ tests/test-fit-oigbeta.R
@@ -10,7 +10,7 @@ set.seed(12345)
 x <- roigbeta(n, 3, 2, 5/2, 1/6)
 
 
-f1 <- fitDR(x, "oigbeta", method="mle", control=list(trace=1, maxit=500)) #
+f1 <- fitDR(x, "oigbeta", method="mle", control=list(trace=0, maxit=500)) 
 summary(f1)
 
 if(FALSE)
@@ -19,7 +19,20 @@ if(FALSE)
 b1 <- bootDR(f1, niter=nboot, silent=TRUE)
 summary(b1)
 
-plot(b1, enhance=TRUE, trueval=c(3, 2, 5/2, 1/6))
+if(sum(b1$converg == 0) > 2)
+{
+  plot(b1, enhance=TRUE, trueval=c(3, 2, 5/2, 1/6))
+  plot(density(b1))
+}
+
+if(FALSE)
+{
+  system.time(b1 <- bootDR(f1, niter=100, silent=TRUE, parallel="snow", ncpus=4))
+  summary(b1)
+  table(b1$converg)
+  plot(density(b1))
+}
+
 
 f2 <- fitDR(x, "oigbeta", method="tlmme")
 summary(f2)