+++ tests/test-fit-oistpareto.R
@@ -1,12 +1,12 @@
 library(mbbefd)
 library(fitdistrplus)
+set.seed(123456)
 
 
 #oistpareto
 n <- 1e3
 nboot <- 1000
 nboot <- 10
-set.seed(12345)
 x <- roistpareto(n, 2, 1/6)
 
 f1 <- fitDR(x, "oistpareto", method="mle", control=list(trace=1))
@@ -19,10 +19,14 @@ if(FALSE)
 b1 <- bootDR(f1, niter=nboot)
 summary(b1)
 
-plot(b1, enhance=TRUE, trueval=c(2, 1/6))
+if(sum(b1$converg == 0) > 2)
+{
+  plot(b1, enhance=TRUE, trueval=c(2, 1/6))
 
-hist(b1$estim[,1])
-hist(b1$estim[,2])
+  hist(b1$estim[,1])
+  hist(b1$estim[,2])
+  plot(density(b1))
+}
 
 f2 <- fitDR(x, "oistpareto", method="tlmme")
 summary(f2)