+++ tests/test-fit-oiunif.R
@@ -1,13 +1,13 @@
 library(mbbefd)
 library(fitdistrplus)
-
+set.seed(123456)
 
 #oiunif
 n <- 1e3
 nboot <- 1000
 nboot <- 10
 x <- roiunif(n, 1/6)
-f1 <- fitDR(x, "oiunif", method="mle", control=list(trace=1))
+f1 <- fitDR(x, "oiunif", method="mle", control=list(trace=0))
 summary(f1)
 
 if(FALSE)
@@ -18,12 +18,16 @@ if(FALSE)
 b1 <- bootDR(f1, niter=nboot)
 summary(b1)
 
-plot(b1)
-abline(v=1/6, col="red")
-
-hist(b1$estim[,1])
-abline(v=1/6, col="red")
-
+if(sum(b1$converg == 0) > 2)
+{
+  plot(b1)
+  abline(v=1/6, col="red")
+  
+  plot(density(b1))
+  
+  hist(b1$estim[,1])
+  abline(v=1/6, col="red")
+}
 
 f2 <- fitDR(x, "oiunif", method="tlmme")