+++ inst/NEWS.md
@@ -4,6 +4,8 @@
 
   * bug fix issue in `fitDR()` for distribution `oibeta` with total-loss-moment-matching estimation.
   * add a `control` argument for `fitDR()` which allows different levels of estimation traces.
+  * add components for `fitDR()` outputs in order to make correct bootstrap estimations in `bootDR()`.
+  * allow parallel computation `bootDR()` as in `bootdist()`.  
 
 # `mbbefd` NEWS